moved test code to test folder
