+++ src/test/integration/scheduler/test_cycle_with_tags.py
@@ -4,14 +4,13 @@ from multiprocessing import Event
 from tempfile import TemporaryDirectory
 from time import sleep
 
-from helperFunctions.fileSystem import get_test_data_dir
 from objects.firmware import Firmware
 from scheduler.Analysis import AnalysisScheduler
-from scheduler.Unpacking import UnpackingScheduler
 from scheduler.analysis_tag import TaggingDaemon
-from storage.MongoMgr import MongoMgr
+from scheduler.Unpacking import UnpackingScheduler
 from storage.db_interface_backend import BackEndDbInterface
-from test.common_helper import get_database_names, clean_test_database
+from storage.MongoMgr import MongoMgr
+from test.common_helper import clean_test_database, get_database_names, get_test_data_dir
 from test.integration.common import initialize_config
