+++ src/test/integration/scheduler/test_regression_virtual_file_path.py
@@ -1,17 +1,17 @@
 from multiprocessing import Event, Value
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from test.common_helper import clean_test_database, get_database_names
-from test.integration.common import initialize_config
 
 import pytest
-from helperFunctions.fileSystem import get_test_data_dir
+
 from intercom.back_end_binding import InterComBackEndBinding
 from objects.firmware import Firmware
 from scheduler.Analysis import AnalysisScheduler
 from scheduler.Unpacking import UnpackingScheduler
 from storage.db_interface_backend import BackEndDbInterface
 from storage.MongoMgr import MongoMgr
+from test.common_helper import clean_test_database, get_database_names, get_test_data_dir
+from test.integration.common import initialize_config
 from web_interface.frontend_main import WebFrontEnd
 
 first_root_id = '5fadb36c49961981f8d87cc21fc6df73a1b90aa1857621f2405d317afb994b64_68415'
