+++ src/test/integration/scheduler/test_unpack_and_analyse.py
@@ -1,14 +1,13 @@
 import gc
-from multiprocessing import Queue
 import unittest
+from multiprocessing import Queue
 from unittest.mock import patch
 
-from helperFunctions.fileSystem import get_test_data_dir
 from objects.firmware import Firmware
 from scheduler.Analysis import AnalysisScheduler
 from scheduler.Unpacking import UnpackingScheduler
-from test.common_helper import DatabaseMock, fake_exit
-from test.integration.common import initialize_config, MockDbInterface, MockFSOrganizer
+from test.common_helper import DatabaseMock, fake_exit, get_test_data_dir
+from test.integration.common import MockDbInterface, MockFSOrganizer, initialize_config
 
 
 class TestFileAddition(unittest.TestCase):
