+++ src/test/integration/statistic/test_work_load.py
@@ -1,12 +1,11 @@
 import gc
-from time import time
 import unittest
+from time import time
 
-from helperFunctions.config import get_config_for_testing
 from statistic.work_load import WorkLoadStatistic
-from storage.MongoMgr import MongoMgr
 from storage.db_interface_statistic import StatisticDbViewer
-from test.common_helper import get_database_names, clean_test_database
+from storage.MongoMgr import MongoMgr
+from test.common_helper import clean_test_database, get_config_for_testing, get_database_names
 
 
 class TestWorkloadStatistic(unittest.TestCase):
