+++ src/test/integration/storage/test_db_interface_backend.py
@@ -1,16 +1,13 @@
 import gc
+import unittest
 from os import path
 from tempfile import TemporaryDirectory
 from time import time
-import unittest
 
-from helperFunctions.config import get_config_for_testing
-from helperFunctions.fileSystem import get_test_data_dir
-from storage.MongoMgr import MongoMgr
 from storage.db_interface_backend import BackEndDbInterface
 from storage.db_interface_common import MongoInterfaceCommon
-from test.common_helper import create_test_firmware, create_test_file_object
-
+from storage.MongoMgr import MongoMgr
+from test.common_helper import create_test_file_object, create_test_firmware, get_config_for_testing, get_test_data_dir
 
 TESTS_DIR = get_test_data_dir()
 test_file_one = path.join(TESTS_DIR, 'get_files_test/testfile1')
