+++ src/test/integration/web_interface/rest/base.py
@@ -2,9 +2,8 @@ import gc
 import unittest
 from tempfile import TemporaryDirectory
 
-from helperFunctions.config import get_config_for_testing
 from storage.MongoMgr import MongoMgr
-from test.common_helper import get_database_names, clean_test_database
+from test.common_helper import clean_test_database, get_config_for_testing, get_database_names
 from web_interface.frontend_main import WebFrontEnd
 
 TMP_DIR = TemporaryDirectory(prefix='fact_test_')
