+++ src/test/unit/analysis/test_yara_plugin_base.py
@@ -1,8 +1,9 @@
 import os
 
 from analysis.YaraPluginBase import YaraBasePlugin
-from helperFunctions.fileSystem import get_src_dir, get_test_data_dir
+from helperFunctions.fileSystem import get_src_dir
 from objects.file import FileObject
+from test.common_helper import get_test_data_dir
 from test.unit.analysis.analysis_plugin_test_class import AnalysisPluginTest
