+++ src/test/unit/helperFunctions/test_config.py
@@ -2,8 +2,8 @@ import os
 
 import pytest
 
-from helperFunctions.fileSystem import get_test_data_dir
 from helperFunctions.config import get_config_dir, load_config, read_list_from_config
+from test.common_helper import get_test_data_dir
 
 
 def test_get_config_dir():
