+++ src/test/unit/helperFunctions/test_hash.py
@@ -1,10 +1,9 @@
 from pathlib import Path
 
-from helperFunctions.fileSystem import get_test_data_dir
 from helperFunctions.hash import (
     check_similarity_of_sets, get_imphash, get_md5, get_sha256, get_ssdeep, get_ssdeep_comparison, normalize_lief_items
 )
-from test.common_helper import create_test_file_object
+from test.common_helper import create_test_file_object, get_test_data_dir
 
 TEST_STRING = 'test string'
 TEST_SHA256 = 'd5579c46dfcc7f18207013e65b44e4cb4e2c2298f4ac457ba8f82743f31e930b'
