+++ src/test/unit/helperFunctions/test_program_setup.py
@@ -1,11 +1,11 @@
-import pytest
 import logging
 import os
 from tempfile import TemporaryDirectory
 
-from helperFunctions.fileSystem import get_test_data_dir
-from helperFunctions.program_setup import _get_console_output_level, _load_config, _setup_logging,\
-    program_setup
+import pytest
+
+from helperFunctions.program_setup import _get_console_output_level, _load_config, _setup_logging, program_setup
+from test.common_helper import get_test_data_dir
 
 
 class argument_mock():
