+++ src/test/unit/objects/test_file.py
@@ -2,8 +2,8 @@ import logging
 
 from common_helper_files import get_binary_from_file
 
-from helperFunctions.fileSystem import get_test_data_dir
 from objects.file import FileObject
+from test.common_helper import get_test_data_dir
 
 
 class TestObjectsFile:  # pylint: disable=no-self-use
