+++ src/test/unit/unpacker/test_tar_repack.py
@@ -1,11 +1,11 @@
 import gc
-import magic
 import os
 import unittest
 
-from helperFunctions.fileSystem import get_test_data_dir
+import magic
+
+from test.common_helper import get_config_for_testing, get_test_data_dir
 from unpacker.tar_repack import TarRepack
-from helperFunctions.config import get_config_for_testing
 
 
 class Test_unpacker_tar_repack(unittest.TestCase):
