+++ src/test/unit/web_interface/base.py
@@ -3,8 +3,7 @@ import unittest
 import unittest.mock
 from tempfile import TemporaryDirectory
 
-from helperFunctions.config import get_config_for_testing
-from test.common_helper import DatabaseMock, fake_exit
+from test.common_helper import DatabaseMock, fake_exit, get_config_for_testing
 from web_interface.frontend_main import WebFrontEnd
 
 TMP_DIR = TemporaryDirectory(prefix="fact_test_")
